+++ apps/frontend/src/graphql/species.graphql.ts
@@ -337,6 +337,7 @@ export const TRAIT_FRAGMENT = gql`
     name
     valueType
     allowsMultipleValues
+    allowsClarifier
     speciesId
     colorId
     color {